Return to reconnect model, fix disappearing status
- The plan eventually is to move to a constantly connected client which
can monitor the presence of the user to avoid marking them as "active"
when they are not. The point is to avoid interfering with mobile push
notifications. However, we aren't there yet. So for now, a reconnect
model seems more reliable.
- This also fixes the disappearing status update by changing the logic
that determines whether we set a status. Instead of checking to see if
the current song matches the last status update, we look at the date
field to determine if the song is currently playing or not. If it is
playing, we update every time. If it is not, we don't. The status should
time out on its own after a couple of minutes.